Shibboleth SSO integration.
With Shibboleth enabled, the login screen in Gitlab is hidden.
Session is started automatically because the user can only access the site after successful login on Single Sign On.

Hide HTTP repository clone URL, when the application is behind Shibboleth access control, the HTTP Basic Auth will not work and therefore HTTP access to repositories is not authenticable.